+++ src/app/modeling/analysis/threat-sources/threat-sources.component.ts
@@ -32,7 +32,9 @@ export class ThreatSourcesComponent implements OnInit {
 
   public AddExistingSource(src: ThreatActor) {
     const actor = this.AddSource();
+    const num = actor.Number;
     actor.CopyFrom(src.Data);
+    actor.Number = num;
     actor.Data['origID'] = src.ID;
   }